Fix set bug where variable was not set with redirection
